# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /inventory/products:
    get:
      operationId: listProducts
      tags: [inventory]
      summary: Lists products and current stock
      responses:
        "200":
          description: Product inventory
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [inventory]
      summary: Creates a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400": { description: Invalid product }
  /inventory/products/{id}:
    patch:
      operationId: updateProduct
      tags: [inventory]
      summary: Updates a product and its stock
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404": { description: Product not found }
    delete:
      operationId: deleteProduct
      tags: [inventory]
      summary: Deletes a product that has no sales
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204": { description: Product deleted }
        "409": { description: Product has sales and cannot be deleted }
        "404": { description: Product not found }
...
  /inventory/purchases/analyze:
    post:
      operationId: analyzePurchaseUpload
      tags: [inventory]
      summary: Extracts a transient spreadsheet, CSV, or image purchase file into an editable preview
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseUploadInput"
      responses:
        "200":
          description: Editable purchase preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchasePreview"
        "400": { description: Invalid or unreadable upload }
        "401": { description: Authentication required }
        "413": { description: Upload too large }
        "415": { description: Unsupported upload type }
        "429": { description: Too many upload analysis requests }
  /inventory/purchases:
    get:
      operationId: listPurchases
      tags: [inventory]
      summary: Lists recorded purchase history
      responses:
        "200":
          description: Purchase history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Purchase"
        "401": { description: Authentication required }
    post:
      operationId: confirmPurchaseImport
      tags: [inventory]
      summary: Records an approved purchase and updates inventory atomically
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseImportInput"
      responses:
        "201":
          description: Recorded purchase
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"
        "400": { description: Invalid purchase }
        "401": { description: Authentication required }
  /sales:
    get:
      operationId: listSales
      tags: [sales]
      summary: Lists recent sales
      responses:
        "200":
          description: Recent sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
    post:
      operationId: createSale
      tags: [sales]
      summary: Registers a sale and discounts stock atomically
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaleInput"
      responses:
        "201":
          description: Registered sale
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
        "400": { description: Invalid sale or insufficient stock }
...
components:
  schemas:
    Product:
      type: object
      properties:
        id: { type: string }
        sku: { type: string }
        name: { type: string }
        category: { type: string }
        description: { type: string }
        cost: { type: number }
        price: { type: number }
        stock: { type: number, minimum: 0 }
        lowStockThreshold: { type: number, minimum: 0 }
        imagePath: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
      required: [id, sku, name, category, description, cost, price, stock, lowStockThreshold, imagePath, createdAt, updatedAt]
    ProductInput:
      type: object
      properties:
        sku: { type: string, minLength: 1, maxLength: 40 }
        name: { type: string, minLength: 1, maxLength: 120 }
        category: { type: string, minLength: 1, maxLength: 80 }
        description: { type: string, maxLength: 500 }
        cost: { type: number, minimum: 0 }
        price: { type: number, minimum: 0 }
        stock: { type: number, minimum: 0 }
        lowStockThreshold: { type: number, minimum: 0 }
        imagePath: { type: ["string", "null"] }
      required: [sku, name, category, description, cost, price, stock, lowStockThreshold, imagePath]
    ProductUpdate:
      type: object
      properties:
        sku: { type: string, minLength: 1, maxLength: 40 }
        name: { type: string, minLength: 1, maxLength: 120 }
        category: { type: string, minLength: 1, maxLength: 80 }
        description: { type: string, maxLength: 500 }
        cost: { type: number, minimum: 0 }
        price: { type: number, minimum: 0 }
        stock: { type: number, minimum: 0 }
        lowStockThreshold: { type: number, minimum: 0 }
        imagePath: { type: ["string", "null"] }
...
    SaleItemInput:
      type: object
      properties:
        productId: { type: string }
        quantity: { type: number, minimum: 1 }
      required: [productId, quantity]
    SaleInput:
      type: object
      properties:
        paymentMethod:
          type: string
          enum: [Efectivo, Tarjeta, Transferencia, Pago móvil]
        amountReceived: { type: ["number", "null"], minimum: 0 }
        customerName: { type: string, minLength: 1, maxLength: 120 }
        customerIdNumber: { type: string, minLength: 1, maxLength: 40 }
        customerPhone: { type: string, minLength: 1, maxLength: 40 }
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/SaleItemInput"
      required: [paymentMethod, amountReceived, customerName, customerIdNumber, customerPhone, items]
    SaleItem:
      type: object
      properties:
        id: { type: string }
        productId: { type: string }
        productName: { type: string }
        quantity: { type: number }
        unitPrice: { type: number }
        unitCost: { type: number }
        subtotal: { type: number }
      required: [id, productId, productName, quantity, unitPrice, unitCost, subtotal]
    Sale:
      type: object
      properties:
        id: { type: string }
        receiptNumber: { type: string }
        paymentMethod: { type: string }
        customerName: { type: string }
        customerIdNumber: { type: string }
        customerPhone: { type: string }
        subtotal: { type: number }
        total: { type: number }
        amountReceived: { type: ["number", "null"] }
        changeDue: { type: number }
        createdAt: { type: string }
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"
      required: [id, receiptNumber, paymentMethod, customerName, customerIdNumber, customerPhone, subtotal, total, amountReceived, changeDue, createdAt, items]
...
    PurchaseUploadInput:
      type: object
      properties:
        fileName: { type: string, minLength: 1, maxLength: 180 }
        mimeType: { type: string, minLength: 1, maxLength: 120 }
        dataBase64: { type: string, minLength: 1, maxLength: 14000000 }
      required: [fileName, mimeType, dataBase64]
    PurchasePreviewItem:
      type: object
      properties:
        sku: { type: string, minLength: 1, maxLength: 40 }
        name: { type: string, minLength: 1, maxLength: 120 }
        category: { type: string, minLength: 1, maxLength: 80 }
        quantity: { type: number, minimum: 1 }
        cost: { type: number, minimum: 0 }
        price: { type: number, minimum: 0 }
        updateCatalog: { type: boolean }
        existingProductId: { type: ["string", "null"] }
        status: { type: string, enum: [existing, new] }
        missingFields:
          type: array
          items: { type: string, enum: [cost, price] }
        warnings:
          type: array
          items: { type: string }
      required: [sku, name, category, quantity, cost, price, updateCatalog, existingProductId, status, missingFields, warnings]
    PurchasePreview:
      type: object
      properties:
        fileName: { type: string }
        items:
          type: array
          items: { $ref: "#/components/schemas/PurchasePreviewItem" }
        warnings:
          type: array
          items: { type: string }
      required: [fileName, items, warnings]
    PurchaseImportItem:
      type: object
      properties:
        sku: { type: string, minLength: 1, maxLength: 40 }
        name: { type: string, minLength: 1, maxLength: 120 }
        category: { type: string, minLength: 1, maxLength: 80 }
        quantity: { type: number, minimum: 1 }
        cost: { type: number, minimum: 0 }
        price: { type: number, minimum: 0 }
        updateCatalog: { type: boolean }
      required: [sku, name, category, quantity, cost, price, updateCatalog]
    PurchaseImportInput:
      type: object
      properties:
        sourceFileName: { type: string, minLength: 1, maxLength: 180 }
        items:
          type: array
          minItems: 1
          maxItems: 2000
          items: { $ref: "#/components/schemas/PurchaseImportItem" }
      required: [sourceFileName, items]
    PurchaseItem:
      type: object
      properties:
        id: { type: string }
        productId: { type: string }
        productName: { type: string }
        sku: { type: string }
        category: { type: string }
        quantity: { type: number }
        unitCost: { type: number }
        unitPrice: { type: number }
        subtotal: { type: number }
      required: [id, productId, productName, sku, category, quantity, unitCost, unitPrice, subtotal]
    Purchase:
      type: object
      properties:
        id: { type: string }
        sourceFileName: { type: string }
        totalCost: { type: number }
        createdAt: { type: string }
        items:
          type: array
          items: { $ref: "#/components/schemas/PurchaseItem" }
      required: [id, sourceFileName, totalCost, createdAt, items]
...