# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: string;
  sku: string;
  name: string;
  category: string;
  description: string;
  cost: number;
  price: number;
  /** @minimum 0 */
  stock: number;
  /** @minimum 0 */
  lowStockThreshold: number;
  /** @nullable */
  imagePath: string | null;
  createdAt: string;
  updatedAt: string;
}