# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  /**
     * @minLength 1
     * @maxLength 40
     */
  sku: string;
  /**
     * @minLength 1
     * @maxLength 120
     */
  name: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  category: string;
  /** @maxLength 500 */
  description: string;
  /** @minimum 0 */
  cost: number;
  /** @minimum 0 */
  price: number;
  /** @minimum 0 */
  stock: number;
  /** @minimum 0 */
  lowStockThreshold: number;
  /** @nullable */
  imagePath: string | null;
}